requirements.txt
Werkzeug==2.2.2
Flask==2.0.3
PyJWT==2.4.0
requests==2.27.1
kubernetes==28.1.0